Updated install to .NET 4.0 Updated the necessary solution items to support new DLLs for service core
git-svn-id: https://fisheye.marc-hi.ca/svn/visualizer/trunk@12 7255dbe7-942c-f64a-b339-be3b494585a5

+++ MARC.EHRS.Visualization.Core/VisualizationEvent.cs
@@ -40,7 +40,7 @@ namespace MARC.EHRS.Visualization.Core
         /// The sequence of the log item
         /// </summary>
         [XmlAttribute("sequence")]
-        public long Sequence { get; set; }
+        public decimal Sequence { get; set; }
 
         /// <summary>
         /// Gets or sets a custom representation of the image
